+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased
 ----------
+- Converted `api` functionality to use `async`/`await` syntax
 - Bumped `test-env-log` dependency to `0.2`